Added many files pretaining to the QueryExec class
Added the general structure for the different types of queries.
The structure for the HistogramExec has been flushed out and can be run
now with a given query string.

In order to make all of this compile, you need to have the Apache Http
Client (http://hc.apache.org/downloads.cgi) and the JSON to Java library,
GSON (http://code.google.com/p/google-gson/downloads/list) on the
source path.