Created the Tweet object in JSON format.
I added tons and tons of files to represent how Twitter represents a Tweet
object in JSON format. It's likely that we won't use the vast majority of it, but
I was unsure of whether or not everything had to be there in order for the GSON
library to work.

This code hasn't been tested to see if it works converting a JSON Tweet into
a Java object yet.