Changed QueryExec objects to extend Thread
Now the HistogramExec, FrequencyExec, and SentimentExec classes
all extend Thread so they can run in parallel. Each class now
has a toString implemented so that they generate a JSON
representation of the data. The TwackerModel now has a toString
method that creates a JSON representation for the whole class
with each of the QueryExec classes' data as well.